[NO-JIRA] Refactor shared helpers
