+++ helpers/common/HelperFunctions.js
