+++ helpers/common/objects.js
