+++ helpers/common/tests/HelperFunctions.test.js
