+++ src/components/common/Card/Card.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { Card as GGCard } from 'gg-components/Card';
-import HelperFunctions from 'helpers/HelperFunctions';
+import HelperFunctions from 'helpers/common/HelperFunctions';
 import Router, { useRouter } from 'next/router';
 import nextifyHref from 'utils/nextifyHref';
