+++ src/components/common/FeatureCard/FeatureCard.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { FeatureCard as GGFeatureCard, FEATURE_CARD_LAYOUTS } from 'gg-components/FeatureCard';
-import HelperFunctions from 'helpers/HelperFunctions';
+import HelperFunctions from 'helpers/common/HelperFunctions';
 import Router, { useRouter } from 'next/router';
 import nextifyHref from 'utils/nextifyHref';
